feat(pod-config)!: Rename `GET /v1/pod/config/dashboard/url` to `GET /v1/dashboard/config/url`
