+++ docs/openapi/openapi.yaml
@@ -97,8 +97,12 @@ paths:
     get: { $ref: "features/pod-config.yaml#/paths/get_pod_config" }
   "/v1/pod/config/address":
     get: { $ref: "features/pod-config.yaml#/paths/get_pod_address" }
-  "/v1/pod/config/dashboard/url":
-    get: { $ref: "features/pod-config.yaml#/paths/get_dashboard_url" }
+
+  # ===== Dashboard config =====
+  "/v1/dashboard/config":
+    get: { $ref: "features/dashboard-config.yaml#/paths/get_dashboard_config" }
+  "/v1/dashboard/config/url":
+    get: { $ref: "features/dashboard-config.yaml#/paths/get_dashboard_url" }
 
   # ===== Workspace =====
   "/v1/workspace":
