+++ src/rest-api/src/features/pod_config/mod.rs
@@ -24,10 +24,6 @@ pub(super) fn router(app_state: AppState) -> axum::Router {
             "/v1/pod/config/address",
             MethodRouter::new().get(get_pod_address_route),
         )
-        .route(
-            "/v1/pod/config/dashboard/url",
-            MethodRouter::new().get(get_dashboard_url_route),
-        )
         .route_layer(from_extractor_with_state::<IsAdmin, _>(app_state.clone()))
         .with_state(app_state)
 }
