+++ src/rest-api/src/features/pod_config/routes.rs
@@ -4,10 +4,7 @@
 // License: Mozilla Public License v2.0 (MPL v2.0)
 
 use axum::{extract::State, Json};
-use service::{
-    app_config::{PodAddress, PodConfig},
-    models::Url,
-};
+use service::app_config::{PodAddress, PodConfig};
 
 use crate::AppState;
 
@@ -21,10 +18,4 @@ pub(super) async fn get_pod_address_route(
     State(AppState { app_config, .. }): State<AppState>,
 ) -> Json<PodAddress> {
     Json(app_config.read().unwrap().pod.clone().address)
-}
-
-pub(super) async fn get_dashboard_url_route(
-    State(AppState { app_config, .. }): State<AppState>,
-) -> Json<Url> {
-    Json(app_config.read().unwrap().dashboard_url().to_owned())
 }
