+++ tests/integration/step-ci/dashboard-config.yaml
@@ -1,0 +1,33 @@
+version: "1.1"
+name: Prose Pod Dashboard configuration
+config:
+  http:
+    baseURL: ${{ env.host }}
+env:
+  $ref: "env.yaml#/env"
+
+# NOTE: Paths are relative to the directory from which `stepci` is ran.
+before:
+  name: "Before all (not a test)"
+  steps:
+    - $ref: "init.yaml#/components/steps/log_admin_in"
+
+tests:
+  getDashboardConfig:
+    # NOTE: Getting the default Dashboard configuration ensures all keys are serialized correctly.
+    #   No need to check _every_ route since we use `serde` in a symetric manner.
+    #   This helps catch most mistakes already.
+    name: Get Dashboard config
+    steps:
+      - name: Get Dashboard config
+        http:
+          method: GET
+          url: /v1/dashboard/config
+          auth:
+            bearer:
+              token: ${{ captures.token }}
+          check:
+            status: 200
+            headers:
+              Content-Type: application/json
+            schema: { $ref: openapi.json#/components/schemas/DashboardConfig }
